added postAge type
